Require minimum macOS version of 10.10
Tectonic has an indirect dependency on the Rustix library, which
references some system calls directly without tests for the minimum
MacOS version, causing link failures against 10.9. In principle
Rustix could get better about this, but it seems pretty unlikely for
that to happen, so in the meantime I think we just have to bump
the minimum OS version.